:recycle: Split http_signing into directory module by responsibility
Pure mechanical move of driver/src/http_signing.rs into a directory module; no behavior change.
