+++ driver/src/http_signing/mod.rs
@@ -1,0 +1,13 @@
+mod actor_key;
+mod cavage;
+mod fetch;
+mod signer;
+mod ssrf;
+#[cfg(test)]
+mod tests;
+mod verifier;
+
+#[cfg(any(test, feature = "test-mode"))]
+pub use actor_key::inject_test_actor_key;
+pub use signer::HttpSignerImpl;
+pub use verifier::HttpSignatureVerifierImpl;
